add form-error file and form-input interface
