+++ btr-web/btr-main-app/utils/form-errors.ts
@@ -1,4 +1,44 @@
-export const emptySharesAndVotes = {
-  message: 'Please enter a percentage of shares and/or votes.',
+export const missingSharesAndVotes = {
+  message: 'Please enter a percentage of shares and/or votes111',
   path: ['percentOfShares', 'percentOfVotes']
+}
+
+export const missingControlOfShares = {
+  message: 'Please indicate at least one type of control',
+  path: ['controlOfShares']
+}
+
+export const missingControlOfDirectors = {
+  message: 'Please indicate at least one type of control',
+  path: ['controlOfDirectors']
+}
+
+export const missingBirthDate = {
+  message: 'Please enter a birth date',
+  path: ['birthDate']
+}
+
+export const missingCitizenship = {
+  message: 'Please indicate citizenship/permanent residency',
+  path: ['citizenshipCA']
+}
+
+export const missingOtherCountry = {
+  message: 'Please select at least one country',
+  path: ['citizenshipsExCA']
+}
+
+export const missingTaxNumberInfo = {
+  message: 'Please select a CRA Tax Number option',
+  path: ['taxNumber']
+}
+
+export const missingTaxResidency = {
+  message: 'Please select a Tax Residency option',
+  path: ['isTaxResident']
+}
+
+export const noMissingInfoReason = {
+  message: 'Please indicate steps taken',
+  path: ['stepsTaken']
 }
