+++ btr-web/btr-main-app/utils/validation.ts
@@ -110,48 +110,3 @@ export function validateTaxResidency (formData: FormInputI): boolean {
 export function validateMissingInfoReasonTextarea(formData: FormInputI): boolean {
   return formData.missingInfoReason === undefined || formData.missingInfoReason.length <= 1000
 }
-
-export const missingSharesAndVotes = {
-  message: 'Please enter a percentage of shares and/or votes111',
-  path: ['percentOfShares', 'percentOfVotes']
-}
-
-export const missingControlOfShares = {
-  message: 'Please indicate at least one type of control',
-  path: ['controlOfShares']
-}
-
-export const missingControlOfDirectors = {
-  message: 'Please indicate at least one type of control',
-  path: ['controlOfDirectors']
-}
-
-export const missingBirthDate = {
-  message: 'Please enter a birth date',
-  path: ['birthDate']
-}
-
-export const missingCitizenship = {
-  message: 'Please indicate citizenship/permanent residency',
-  path: ['citizenshipCA']
-}
-
-export const missingOtherCountry = {
-  message: 'Please select at least one country',
-  path: ['citizenshipsExCA']
-}
-
-export const missingTaxNumberInfo = {
-  message: 'Please select a CRA Tax Number option',
-  path: ['taxNumber']
-}
-
-export const missingTaxResidency = {
-  message: 'Please select a Tax Residency option',
-  path: ['isTaxResident']
-}
-
-export const noMissingInfoReason = {
-  message: 'Please indicate steps taken',
-  path: ['stepsTaken']
-}
